browser/requirements.txt
flask==1.1.1
flask-cors==3.0.8
flask-sqlalchemy==2.4.1
matplotlib==3.0.3
mysqlclient==1.4.5
numpy>=1.16.4
boto3==1.9.182
scikit-image>=0.15.0,<0.17.0
python-decouple==3.1
pillow>=7.1.0
flask-compress==1.5.0
flask_monitoringdashboard==3.1.0